Added a new private field lastTrigrams of type int[] to store the last trigram values.
Modified the constructor of the Hexagram class to initialize the lastTrigrams array.
Modified the GetTrigrams() method to check if the _values.Changed property is true. If it is true, it updates the lastTrigrams array.
Updated the UpdateTrigrams() method to use the row parameter instead of the pos parameter.
Modified the UpdateTrigrams() method to update the lastTrigrams array instead of returning an array.